Rename .inc files to .h files to handle build dependencies since changes to .h files already cause rebuilds
Add k == 0 handling

Replace fa4 = 0 hack with more descriptive name fzero which is zero throughout
entire routine
