+++ kernels/rviv/3/bli_sgemm_rviv_asm_4vx4.S
@@ -42,4 +42,4 @@
 #define VLE vle32.v
 #define VSE vse32.v
 
-#include "bli_sdgemm_rviv_asm_4vx4.inc"
+#include "bli_sdgemm_rviv_asm_4vx4.h"
