+++ kernels/rviv/3/bli_zgemm_rviv_asm_4vx4.S
@@ -41,4 +41,4 @@
 #define VLE vlseg2e64.v
 #define VSE vsseg2e64.v
 
-#include "bli_czgemm_rviv_asm_4vx4.inc"
+#include "bli_czgemm_rviv_asm_4vx4.h"
